modify type of indices

+++ driver/dm_gather.cpp
@@ -29,11 +29,11 @@
 static Driver* makeDriver(const std::string& base_arg)
 {
     if(base_arg == "gather")
-        return new GatherDriver<float, float, int>();
+        return new GatherDriver<float, float, uint64_t>();
     if(base_arg == "gatherfp16")
-        return new GatherDriver<float16, float, int>();
+        return new GatherDriver<float16, float, uint64_t>();
     if(base_arg == "gatherbfp16")
-        return new GatherDriver<bfloat16, float, int>();
+        return new GatherDriver<bfloat16, float, uint64_t>();
     return nullptr;
 }
